+++ index.html
@@ -10,6 +10,16 @@
     <!-- AddEvent -->
   </body>
   <!-- <script type="text/javascript" src="https://tool.votinginfoproject.org/app.js"></script> -->
-  <script type="text/javascript" src="/src/components/CalendarImport/stc.min.js"></script>
+  <script type="text/javascript" src="/src/components/CalendarPickerButtons/stc.min.js"></script>
+  <!-- AddEvent Settings -->
+  <script type="text/javascript">
+    window.addeventasync = function(){
+      addeventstc.settings({
+          license    : "auYdxlyGjzxZCenXymdJ25252"
+      });
+    };
+  </script>
+
+
   <script src="/bundle.js"></script>
 </html>
