+++ src/components/CalendarPickerButtons/css/theme3.css
@@ -1,0 +1,63 @@
+.calendarBtnContainer {
+  font-family: Roboto, "Segoe UI", Frutiger, "Frutiger Linotype", "Dejavu Sans", "Helvetica Neue", Arial, sans-serif;
+  display: inline-block;
+  position: relative;
+  text-align: left;
+  z-index: 999998;
+  cursor: pointer;
+  color: #fff!important;
+  font-size: 17px;
+  font-weight: 300;
+  line-height: 100%;
+  width: 100%;
+  text-decoration: none;
+}
+
+.calendarBtnContainer span {
+  display: block;
+  background: #fff;
+  margin-right: 10px;
+  position: relative;
+  line-height: 100%;
+  text-decoration: none;
+  font-size: 14px;
+  color: #333;
+  padding: 9px 10px 9px 40px;
+}
+
+.calendarBtnContainer span:hover {
+    color: #fff;
+    font-size: 17px;
+    text-decoration: none;
+    background-color: #ccc;
+}
+
+.calendarBtnContainer .ateappleical {
+    background-image: url(../gfx/dropdown-apple-t1.png);
+    background-position: 13px 50%;
+    background-repeat: no-repeat;
+}
+
+.calendarBtnContainer .ateoutlook {
+    background-image: url(../gfx/dropdown-outlook-t1.png);
+    background-position: 13px 50%;
+    background-repeat: no-repeat;
+}
+
+.calendarBtnContainer .ateoutlookcom {
+    background-image: url(../gfx/dropdown-outlook-t1.png);
+    background-position: 13px 50%;
+    background-repeat: no-repeat;
+}
+
+.calendarBtnContainer .ategoogle {
+    background-image: url(../gfx/dropdown-google-t1.png);
+    background-position: 13px 50%;
+    background-repeat: no-repeat;
+}
+
+.calendarBtnContainer .ateyahoo {
+    background-image: url(../gfx/dropdown-yahoo-t1.png);
+    background-position: 13px 50%;
+    background-repeat: no-repeat;
+}
